move classes from cloud storage project
